requirements.txt
Flask==3.0.0
pywebview==5.0.5
pystray==0.19.5
Pillow==10.2.0
psutil==5.9.8
pywin32==306